# Swagger_json/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "v1",
    "title": "My API"
  },
  "paths": {
...
    "/api/Reservas": {
      "get": {
        "tags": [
          "Reservas"
        ],
        "operationId": "GetReservas",
        "consumes": [],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "uniqueItems": false,
              "type": "array",
              "items": {
                "$ref": "#/definitions/Reservas"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Reservas"
        ],
        "operationId": "PostReservas",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [],
        "parameters": [
          {
            "name": "reservas",
            "in": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/Reservas"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Reservas/{id}": {
      "get": {
        "tags": [
          "Reservas"
        ],
        "operationId": "GetReservas",
        "consumes": [],
        "produces": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Reservas"
        ],
        "operationId": "PutReservas",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "reservas",
            "in": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/Reservas"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Reservas"
        ],
        "operationId": "DeleteReservas",
        "consumes": [],
        "produces": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Reservas/{p_date}": {
      "post": {
        "tags": [
          "Reservas"
        ],
        "operationId": "DameReservas",
        "consumes": [],
        "produces": [],
        "parameters": [
          {
            "name": "p_date",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "date-time"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Socios": {
      "get": {
        "tags": [
          "Socios"
        ],
        "operationId": "GetSocios",
        "consumes": [],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "uniqueItems": false,
              "type": "array",
              "items": {
                "$ref": "#/definitions/Socios"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Socios"
        ],
        "operationId": "PostSocios",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [],
        "parameters": [
          {
            "name": "socios",
            "in": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/Socios"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Socios/{id}": {
      "get": {
        "tags": [
          "Socios"
        ],
        "operationId": "GetSocios",
        "consumes": [],
        "produces": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Socios"
        ],
        "operationId": "PutSocios",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "socios",
            "in": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/Socios"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Socios"
        ],
        "operationId": "DeleteSocios",
        "consumes": [],
        "produces": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Usuarios": {
      "get": {
        "tags": [
          "Usuarios"
        ],
        "operationId": "GetUsuarios",
        "consumes": [],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "uniqueItems": false,
              "type": "array",
              "items": {
                "$ref": "#/definitions/Usuarios"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Usuarios"
        ],
        "operationId": "PostUsuarios",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [],
        "parameters": [
          {
            "name": "usuarios",
            "in": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/Usuarios"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Usuarios/{id}": {
      "get": {
        "tags": [
          "Usuarios"
        ],
        "operationId": "GetUsuarios",
        "consumes": [],
        "produces": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Usuarios"
        ],
        "operationId": "PutUsuarios",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "usuarios",
            "in": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/Usuarios"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Usuarios"
        ],
        "operationId": "DeleteUsuarios",
        "consumes": [],
        "produces": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
...
    "Socios": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "surname": {
          "type": "string"
        },
        "telephone": {
          "format": "int32",
          "type": "integer"
        },
        "dni": {
          "type": "string"
        },
        "user": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "Reservas": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "dni": {
          "type": "string"
        },
        "sport": {
          "type": "string"
        },
        "nField": {
          "format": "int32",
          "type": "integer"
        },
        "date": {
          "format": "date-time",
          "type": "string"
        }
      }
    },
    "Usuarios": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "user": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    }
  }
}